Take advantage of knowing what pins you picked.
Since the i2c code has hardcoded values for _data_pins[x],
which are set to fixed values in the only case where _i2caddr
isn't 255, why not just take advantage of the fact that we
want to put four consecutive bits in four consecutive bits,
and we know which bits in advance?

No idea of performance impact, and instructions are pretty
cheap, but this removes a lot of shifts which necessarily have
to be done at runtime (because the compiler can't tell that
these values are actually fixed), plus some looping and array
indexing.